Move state tests
