+++ packages/state/src/VaultLeaf.test.ts
@@ -1,7 +1,7 @@
 import { PedersenHash, StarkKey } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { VaultLeaf } from '../src/VaultLeaf'
+import { VaultLeaf } from './VaultLeaf'
 
 const tokenA = PedersenHash(
   '0xd5b742d29ab21fdb06ac5c7c460550131c0b30cbc4c911985174c0ea4a92ec'
